+++ src/types/mod.rs
@@ -1,4 +1,5 @@
 pub mod attr;
 pub mod file;
+pub mod handle;
 pub mod inode;
 pub mod torrent;